[powerpc] kvm: Add provisioning for ulong vm stats and u64 vcpu stats
Message-id: <1484749104-27456-5-git-send-email-thuth@redhat.com>
Patchwork-id: 162756
O-Subject: [RHEL7.4 kvm PATCH 04/12] KVM: Add provisioning for ulong vm stats and u64 vcpu stats
Bugzilla: 1399882
RH-Acked-by: David Hildenbrand <david@redhat.com>
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>

From: Suraj Jitindar Singh <sjitindarsingh@gmail.com>

vms and vcpus have statistics associated with them which can be viewed
within the debugfs. Currently it is assumed within the vcpu_stat_get() and
vm_stat_get() functions that all of these statistics are represented as
u32s, however the next patch adds some u64 vcpu statistics.

Change all vcpu statistics to u64 and modify vcpu_stat_get() accordingly.
Since vcpu statistics are per vcpu, they will only be updated by a single
vcpu at a time so this shouldn't present a problem on 32-bit machines
which can't atomically increment 64-bit numbers. However vm statistics
could potentially be updated by multiple vcpus from that vm at a time.
To avoid the overhead of atomics make all vm statistics ulong such that
they are 64-bit on 64-bit systems where they can be atomically incremented
and are 32-bit on 32-bit systems which may not be able to atomically
increment 64-bit numbers. Modify vm_stat_get() to expect ulongs.

Acked-by: Christian Borntraeger <borntraeger@de.ibm.com>
(cherry picked from commit 8a7e75d47b68193339f8727cf4503271d0a0b1d0)

Conflicts:
	arch/arm/include/asm/kvm_host.h
	arch/arm64/include/asm/kvm_host.h
	arch/mips/include/asm/kvm_host.h
	arch/powerpc/include/asm/kvm_host.h
	arch/s390/include/asm/kvm_host.h
	arch/x86/include/asm/kvm_host.h
	(contextual conflicts in struct kvm_vcpu_stat since we do not
	 have commit 3491caf2755e9f312666712510d80b00c81ff247 "KVM:
	 halt_polling: provide a way to qualify wakeups during poll"
	 in downstream yet)

+++ arch/arm/include/asm/kvm_host.h
@@ -141,13 +141,13 @@ struct kvm_vcpu_arch {
 };
 
 struct kvm_vm_stat {
-	u32 remote_tlb_flush;
+	ulong remote_tlb_flush;
 };
 
 struct kvm_vcpu_stat {
-	u32 halt_successful_poll;
-	u32 halt_attempted_poll;
-	u32 halt_wakeup;
+	u64 halt_successful_poll;
+	u64 halt_attempted_poll;
+	u64 halt_wakeup;
 };
 
 int kvm_vcpu_set_target(struct kvm_vcpu *vcpu,
